Remove bound scale property and move some functions from meshhelper into VectorHelper
